テスト用にlimits.hをインクルード

+++ ft_printf/includes/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: kishigam <kishigam@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/25 11:10:28 by kishigam          #+#    #+#             */
-/*   Updated: 2022/06/11 23:06:24 by kishigam         ###   ########.fr       */
+/*   Updated: 2022/06/12 08:37:47 by kishigam         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,6 +17,7 @@
 # include <unistd.h>
 # include <stdarg.h>
 # include <stdio.h>
+# include <limits.h>
 
 int	ft_printf(const char *format, ...);
 int	ft_parse(char c, va_list format);
@@ -25,6 +26,6 @@ int	ft_putstr(char *str);
 int	ft_putptr(void *ptr);
 int	ft_putint(int num);
 int	ft_putuint(unsigned int num);
-int	ft_puthex(size_t num, char c);
+int	ft_puthex(unsigned int num, char c);
 
 #endif
